pyproject.toml
[project]
name = "mcp-diffusion-server-20260415"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastmcp>=3.2.4",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
